refactor: rename CoreSocketConnector to CloudSqlConnectorRegistry.
